look for berkeley-abc in the PATH
Before this change to the patch, yosys would determine the directory
that the yosys program lived in and would look for abc there.  After this
change, it just looks in the PATH.

This change is generally an improvement because it simplifies the code,
but it's also required to build the documentation.

When we're building the package, abc is in /usr/bin/ (provided by the
build-dependency on berkeley-abc).  Before this change to the patch,
the abc.cc code used look for abc in the same directory as the test
program lived in, and thus would fail to find it.